+++ sys/src/libc/port/strncmp.c
@@ -2,7 +2,7 @@
 #include <libc.h>
 
 int
-strncmp(char *s1, char *s2, int n)
+strncmp(char *s1, char *s2, long n)
 {
 	unsigned c1, c2;
